Change pool margin from Double to BigDecimal

+++ calculation/src/main/java/org/cardanofoundation/rewards/calculation/domain/PoolState.java
@@ -2,6 +2,7 @@ package org.cardanofoundation.rewards.calculation.domain;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.HashSet;
 
@@ -17,7 +18,7 @@ public class PoolState {
     private HashSet<String> owners;
     private BigInteger ownerActiveStake;
     private BigInteger poolFees;
-    private Double margin;
+    private BigDecimal margin;
     private BigInteger fixedCost;
     private BigInteger pledge;
     private HashSet<Delegator> delegators;
